// apps/intelli-law/lib/database.types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      conversations: {
        Row: {
          account_id: string
          created_at: string
          document_id: string
          id: number
          name: string
          reference_id: string
        }
        Insert: {
          account_id: string
          created_at?: string
          document_id: string
          id?: number
          name: string
          reference_id: string
        }
        Update: {
          account_id?: string
          created_at?: string
          document_id?: string
          id?: number
          name?: string
          reference_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "conversations_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "conversations_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "user_account_workspace"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "conversations_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "user_accounts"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "conversations_document_id_fkey"
            columns: ["document_id"]
            isOneToOne: false
            referencedRelation: "documents"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
// ...
}